full fix

+++ examples/simulacra.py
@@ -1,6 +1,7 @@
 # Optimize prompts by training on prompts-ratings pairings dataset
 # taken from https://github.com/JD-P/simulacra-aesthetic-captions
 
+from accelerate import Accelerator
 import os
 import sqlite3
 import time
@@ -13,9 +14,11 @@ url = "https://raw.githubusercontent.com/JD-P/simulacra-aesthetic-captions/main/
 dbpath = "sac_public_2022_06_29.sqlite"
 
 if __name__ == "__main__":
-    if not os.path.exists(dbpath):
-        print(f"fetching {dbpath}")
-        urlretrieve(url, dbpath)
+    accelerator = Accelerator()
+    with accelerator.local_main_process_first():
+        if os.environ.get('LOCAL_RANK', 0) == 0 and not os.path.exists(dbpath):
+            print(f"fetching {dbpath}")
+            urlretrieve(url, dbpath)
 
     conn = sqlite3.connect(dbpath)
     c = conn.cursor()
